Authorization/requirements.txt
flask 
flask-sqlalchemy 
flask-login 
flask-wtf 
werkzeug
mysql-connector-python
pymysql
Flask
flask-bcrypt